remove deprecated PoseAtTime and SkycamModel
